+++ public/style.css
@@ -108,3 +108,7 @@ td.column-td {
 .icon {
   font-size: 1.5rem;
 }
+
+.emoji-option {
+  height: 30px;
+}
